@@ -1,19 +1,19 @@
-Rails.application.routes.draw do
-  scope "/#{(Spree::Config[:blog_title] || 'blog').to_url}" do
-    match '/tagged/:tag',           :to => 'posts#index', :as => 'posts_by_tag'
-    match '/:year(/:month(/:day))', :to => 'posts#index', :as => 'posts_by_date', 
-                                                            :year  => /\d{4}/,
-                                                            :month => /\d{1,2}/,
-                                                            :day   => /\d{1,2}/
-    resources :posts, :path => '/' do
+Spree::Core::Engine.routes.prepend do
+  scope "/#{(Spree::Config.blog_title || 'blog').to_url}" do
+    match '/tagged/:tag',           to: 'posts#index', as: 'posts_by_tag'
+    match '/:year(/:month(/:day))', to: 'posts#index', as: 'posts_by_date',
+                                                            year:  /\d{4}/,
+                                                            month: /\d{1,2}/,
+                                                            day:   /\d{1,2}/
+    resources :posts, path: '/' do
       resources :comments
     end
   end
 
-  namespace :admin do 
+  namespace :admin do
     resources :posts do
       resources :comments
-      resources :images, :controller => 'post_images'
+      resources :images, controller: 'post_images'
     end
     resources :comments
   end
