@@ -1,14 +1,30 @@
 module SpreeAwesomeBlog
   module Generators
     class InstallGenerator < Rails::Generators::Base
-      source_root File.expand_path("../../templates", __FILE__)
+      class_option :auto_run_migrations, type: :boolean, default: false
 
-      desc "Configures your Rails application for use with spree_awesome_blog"
-      def copy_files
-        directory "db"
-        directory "public"
+      def add_javascripts
+        append_file 'app/assets/javascripts/store/all.js', "//= require store/spree_awesome_blog\n"
+        append_file 'app/assets/javascripts/admin/all.js', "//= require admin/spree_awesome_blog\n"
       end
 
+      def add_stylesheets
+        inject_into_file 'app/assets/stylesheets/admin/all.css', " *= require admin/spree_awesome_blog\n", before: /\*\//, verbose: true
+        inject_into_file 'app/assets/stylesheets/store/all.css', " *= require store/spree_awesome_blog\n", before: /\*\//, verbose: true
+      end
+
+      def add_migrations
+        run 'bundle exec rake railties:install:migrations FROM=spree_awesome_blog'
+      end
+
+      def run_migrations
+        run_migrations = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask 'Would you like to run the migrations now? [Y/n]')
+        if run_migrations
+          run 'bundle exec rake db:migrate'
+        else
+          puts 'Skipping rake db:migrate, don\'t forget to run it!'
+        end
+      end
     end
   end
 end
