@@ -1,5 +1,5 @@
-require 'spree_core'
-require 'spree_awesome_blog_hooks'
+require 'spree/core'
+require 'spree_awesome_blog/version'
 require 'spree_awesome_blog/engine'
 require 'acts-as-taggable-on'
 require 'rdiscount'
